Move "RequestManager" in sub folder
